pdf_chat_bot/requirement.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
pinecone-client
faiss-cpu
langchain_google_genai
langchain-community
langchain_pinecone